django/requirements.txt
Django==1.6.1
Pillow==2.2.2
argparse==1.2.1
django-admin-sortable==1.6.3
django-ckeditor==4.0.2
django-extensions==1.2.5
django-grappelli==2.5.1
psycopg2==2.5.1
six==1.4.1
wsgiref==0.1.2
